+++ package.json
@@ -25,10 +25,8 @@
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-jest": "^24.1.0",
-    "firebase-mock": "^2.3.2",
     "htmlhint": "^0.14.0",
     "jest": "^27.0.1",
-    "mock-cloud-firestore": "^0.12.0",
     "regenerator-runtime": "^0.13.1",
     "serve": "^12.0.0",
     "stylelint": "^13.13.1",
