+++ src/css/post2.css
@@ -20,16 +20,16 @@ a {
     background: var(--white);
     width: 100%;
     max-width: 60rem;
-    border-radius: 1rem;
+    border-radius: 0.5rem;
     box-shadow: 0 1px 15px 1px rgba(69, 65, 78, .08);
     margin-bottom: 1rem;
 }
 
 .post-header {
     display: flex;
+    border-radius: 0.5rem;
     justify-content: space-between;
     padding: 1.2rem 1.6rem 0 1.6rem;
-    border-radius: 1rem;
 }
 
 
@@ -37,6 +37,11 @@ a {
 
 .header-left {
     flex: 1;
+}
+
+.author-name:hover,
+.post-date:hover {
+    text-decoration: underline;
 }
 
 .author-name {
@@ -71,12 +76,6 @@ a {
     font-size: 12PX;
     color: #878787;
 }
-
-.author-name:hover,
-.post-date:hover {
-    text-decoration: underline;
-}
-
 
 
 /* Dot between post author and post date */
@@ -132,9 +131,8 @@ a {
 .content-image {
     width: 90%;
     height: 342px;
-    /* object-fit: cover; */
+    object-fit: cover;
     display: block;
-    background-size: cover;
     margin: 0 auto;
     border-radius: 10px;
 }
@@ -145,7 +143,7 @@ a {
    ***************/
 
 .footer {
-    padding: .3rem 1rem;
+    padding: 1rem 1.2rem;
 }
 
 
@@ -154,7 +152,7 @@ a {
 .footer-reactions {
     display: flex;
     align-items: center;
-    padding: 0.5rem 1.6rem;
+    padding: 0.8rem 1.6rem;
     color: var(--reactions-links);
     font-size: 1.5rem;
 }
